Double заменен на BigDecimal во внутренних моделях и мапперах. Исправлены тесты

+++ crypto-common/src/commonMain/kotlin/models/CryptoCurrency.kt
@@ -1,6 +1,8 @@
 package models
 
+import java.math.BigDecimal
+
 data class CryptoCurrency(
     var ticker: String = "",
-    var value: Double = 0.0
+    var value: BigDecimal = BigDecimal.valueOf(0.0)
 )
